requirements.txt
icecream==2.1.3
json_tricks==3.16.1
mmcv_full==1.7.0
mmpose.egg==info
numpy==1.21.5
numpy==1.22.4
opencv_python==4.5.5.64
Pillow==9.0.1
Pillow==10.0.0
torch==1.11.0
xtcocotools==1.12
